Adding the possibility to have a configuration file
I didn't want to have to stash my own configuration for every change.

The config file is loaded as a bash file, and can contain values for
the existing configurable properties:

- `snapshot_interval_seconds`
- `snapshots_per_second`
- `keep_video_days`
- `videos_location` (one level below the "captures" folder)

And an additional `show_config` variable which triggers, if set to
something different than 0, the display of a summary of the
configuration values.